feat(phase-04-pr04n): remove migration-window shims; Color type moves to domain
- Adds domain.Color type (uint8 + ColorHealthy/Warning/Broken/Dim constants)
- Adds Color + Augment fields to catalog.ResourceTypeDef with ResolveColor method
- Inlines all per-category color functions into catalog/types_*.go files
- Adds catalog/color_helpers.go with shared helpers (fallback, cfnStack, ACM, R53)
- Deletes all 12 internal/resource/color_*.go migration-window shim files
- Removes colorRegistry, augmentRegistry, projectRegistry, adaptFromCatalog,
  buildResourceTypes, and var resourceTypes from internal/resource/types.go
- ResourceTypeDef and Color become type aliases for catalog/domain counterparts
- AllResourceTypes/AllShortNames/FindResourceType become pure catalog passthroughs
- Adds ct_events_init.go cycle-breaker to wire ctevent.Project into catalog
