+++ internal/catalog/types_dns_cdn.go
@@ -2,7 +2,22 @@ package catalog
 
 import "github.com/k2m30/a9s/v3/internal/domain"
 
-var dnsCdnTypes = []ResourceTypeDef{
+func colorCF(r domain.Resource) domain.Color {
+	if r.Fields["enabled"] == "false" {
+		return domain.ColorDim
+	}
+	switch r.Fields["status"] {
+	case "Deployed":
+		return domain.ColorHealthy
+	case "InProgress":
+		return domain.ColorWarning
+	}
+	return domain.ColorHealthy
+}
+
+func colorAPIGW(_ domain.Resource) domain.Color { return domain.ColorHealthy }
+
+var dnsCdnTypes = []ResourceTypeDef{ //nolint:gochecknoglobals // static catalog: intentional package-level var
 	{
 		Name:          "Route 53 Hosted Zones",
 		ShortName:     "r53",
@@ -22,6 +37,7 @@ var dnsCdnTypes = []ResourceTypeDef{
 			ContextKeys:    map[string]string{"zone_id": "ID", "zone_name": "Name"},
 			DisplayNameKey: "zone_name",
 		}},
+		Color: r53Color,
 	},
 	{
 		Name:          "CloudFront Distributions",
@@ -37,6 +53,7 @@ var dnsCdnTypes = []ResourceTypeDef{
 			{Key: "aliases", Title: "Aliases", Width: 30, Sortable: false},
 			{Key: "price_class", Title: "Price Class", Width: 16, Sortable: true},
 		},
+		Color: colorCF,
 	},
 	{
 		Name:          "ACM Certificates",
@@ -51,6 +68,7 @@ var dnsCdnTypes = []ResourceTypeDef{
 			{Key: "not_after", Title: "Expires", Width: 22, Sortable: true},
 			{Key: "in_use", Title: "In Use", Width: 8, Sortable: true},
 		},
+		Color: acmColor,
 	},
 	{
 		Name:          "API Gateways",
@@ -65,5 +83,6 @@ var dnsCdnTypes = []ResourceTypeDef{
 			{Key: "endpoint", Title: "Endpoint", Width: 50, Sortable: false},
 			{Key: "description", Title: "Description", Width: 30, Sortable: false},
 		},
+		Color: colorAPIGW,
 	},
 }
