+++ internal/catalog/types_messaging.go
@@ -1,8 +1,77 @@
 package catalog
 
-import "github.com/k2m30/a9s/v3/internal/domain"
+import (
+	"strings"
 
-var messagingTypes = []ResourceTypeDef{
+	"github.com/k2m30/a9s/v3/internal/domain"
+)
+
+func colorSQS(_ domain.Resource) domain.Color { return domain.ColorHealthy }
+func colorSNS(_ domain.Resource) domain.Color { return domain.ColorHealthy }
+func colorSFN(_ domain.Resource) domain.Color { return domain.ColorHealthy }
+
+func colorSNSSub(r domain.Resource) domain.Color {
+	switch r.Fields["subscription_arn"] {
+	case "PendingConfirmation":
+		return domain.ColorWarning
+	case "Deleted":
+		return domain.ColorDim
+	default:
+		return domain.ColorHealthy
+	}
+}
+
+func colorEBRule(r domain.Resource) domain.Color {
+	switch strings.ToUpper(r.Fields["state"]) {
+	case "ENABLED", "ENABLED_WITH_ALL_CLOUDTRAIL_MANAGEMENT_EVENTS":
+		return domain.ColorHealthy
+	case "DISABLED":
+		return domain.ColorDim
+	}
+	return domain.ColorHealthy
+}
+
+func colorKinesis(r domain.Resource) domain.Color {
+	switch r.Fields["stream_status"] {
+	case "ACTIVE":
+		return domain.ColorHealthy
+	case "CREATING", "UPDATING", "DELETING":
+		return domain.ColorWarning
+	}
+	switch r.Fields["status"] {
+	case "ACTIVE":
+		return domain.ColorHealthy
+	case "CREATING", "UPDATING", "DELETING":
+		return domain.ColorWarning
+	}
+	return domain.ColorHealthy
+}
+
+func colorMSK(r domain.Resource) domain.Color {
+	switch r.Fields["state"] {
+	case "ACTIVE":
+		return domain.ColorHealthy
+	case "CREATING", "UPDATING", "MAINTENANCE", "REBOOTING_BROKER", "HEALING":
+		return domain.ColorWarning
+	case "FAILED":
+		return domain.ColorBroken
+	}
+	return domain.ColorHealthy
+}
+
+func colorSES(r domain.Resource) domain.Color {
+	phrase := stripFindingSuffix(r.Fields["status"])
+	switch phrase {
+	case "verification failed", "verify: temp failure", "verification not started",
+		"account SHUTDOWN", "account PROBATION":
+		return domain.ColorBroken
+	case "pending verification", "sending disabled":
+		return domain.ColorWarning
+	}
+	return domain.ColorHealthy
+}
+
+var messagingTypes = []ResourceTypeDef{ //nolint:gochecknoglobals // static catalog: intentional package-level var
 	{
 		Name:          "SQS Queues",
 		ShortName:     "sqs",
@@ -16,6 +85,7 @@ var messagingTypes = []ResourceTypeDef{
 			{Key: "delay_seconds", Title: "Delay", Width: 8, Sortable: true},
 			{Key: "queue_url", Title: "Queue URL", Width: 50, Sortable: false},
 		},
+		Color: colorSQS,
 	},
 	{
 		Name:          "SNS Topics",
@@ -33,6 +103,7 @@ var messagingTypes = []ResourceTypeDef{
 			ContextKeys:    map[string]string{"topic_arn": "ID"},
 			DisplayNameKey: "display_name",
 		}},
+		Color: colorSNS,
 	},
 	{
 		Name:          "SNS Subscriptions",
@@ -46,6 +117,7 @@ var messagingTypes = []ResourceTypeDef{
 			{Key: "endpoint", Title: "Endpoint", Width: 48, Sortable: false},
 			{Key: "subscription_arn", Title: "Subscription ARN", Width: 60, Sortable: false},
 		},
+		Color: colorSNSSub,
 	},
 	{
 		Name:          "EventBridge Rules",
@@ -66,6 +138,7 @@ var messagingTypes = []ResourceTypeDef{
 			ContextKeys:    map[string]string{"rule_name": "ID", "event_bus": "event_bus"},
 			DisplayNameKey: "rule_name",
 		}},
+		Color: colorEBRule,
 	},
 	{
 		Name:          "Kinesis Streams",
@@ -79,6 +152,7 @@ var messagingTypes = []ResourceTypeDef{
 			{Key: "stream_mode", Title: "Mode", Width: 14, Sortable: true},
 			{Key: "creation_time", Title: "Created", Width: 22, Sortable: true},
 		},
+		Color: colorKinesis,
 	},
 	{
 		Name:          "MSK Clusters",
@@ -92,6 +166,7 @@ var messagingTypes = []ResourceTypeDef{
 			{Key: "state", Title: "State", Width: 14, Sortable: true},
 			{Key: "version", Title: "Version", Width: 14, Sortable: true},
 		},
+		Color: colorMSK,
 	},
 	{
 		Name:          "Step Functions",
@@ -115,6 +190,7 @@ var messagingTypes = []ResourceTypeDef{
 			},
 			DrillBlockMessage: "Execution history is not available for Express state machines",
 		}},
+		Color: colorSFN,
 	},
 	{
 		Name:          "SES Identities",
@@ -127,5 +203,6 @@ var messagingTypes = []ResourceTypeDef{
 			{Key: "identity_type", Title: "Type", Width: 16, Sortable: true},
 			{Key: "status", Title: "Status", Width: 36, Sortable: true},
 		},
+		Color: colorSES,
 	},
 }
